Rename publishing containers (#117)
